refactor : Change component name Map to Main
