+++ src/App.tsx
@@ -1,12 +1,14 @@
 import React from 'react'
 import { Route, Routes } from 'react-router-dom'
-import Map from '@pages/Map/Map'
+import Main from '@pages/main/Main'
+import Review from '@pages/review/Review'
 
 function App() {
   return (
     <div>
       <Routes>
-        <Route path="/" element={<Map />} />
+        <Route path="/" element={<Main />} />
+        <Route path="/review" element={<Review />} />
       </Routes>
     </div>
   )
