+++ src/pages/review/Review.tsx
@@ -1,0 +1,7 @@
+import React from 'react'
+
+const Review = () => {
+  return <div>Review</div>
+}
+
+export default Review
